+++ classes/prefix_WPgutenberg/blocks/core-extensions/video/attributes.js
@@ -1,0 +1,12 @@
+const attributes = {
+  videoJS: {
+    type: 'boolean',
+    default: false,
+  },
+  playCentered: {
+    type: 'boolean',
+    default: false,
+  }
+}
+
+export default attributes;